refactor : 유틸 함수 생성 추가, 삭제, 수정 시 토글 유지하는 함수
