+++ README.md
@@ -20,4 +20,5 @@
 - [x] 루트 페이지 추가 기능
 - [x] 단일 페이지 삭제 기능
 - [x] 하위페이지 토글 기능
+- [x] 추가, 수정, 삭제 시에도 펼침기능 유지
 - [ ] CSS 작업
