+++ src/App.js
@@ -2,6 +2,7 @@ import Button from "./Button";
 import DocumentContent from "./DocumentContent";
 import DocumentList from "./DocumentList";
 import { request } from "./api.js";
+import { getRightNextState } from "./util";
 
 export default class App {
   constructor() {
@@ -17,38 +18,11 @@ export default class App {
         const { id } = await request.addDocumentItem(0);
         history.pushState(null, null, `/${id}`);
         this.documentContent.setState(await request.getDocumentItem(id));
-        // 현재 데이터에서 isSpread가 true인 id를 results 배열에 넣는다.
-        const prevState = structuredClone(this.documentList.state);
 
-        const stack = prevState;
-        const results = [];
-        // 트리 순회
-        while (stack.length) {
-          const currentNode = stack.pop();
-
-          if (currentNode.isSpread) {
-            results.push(currentNode.id);
-          }
-
-          for (let i = 0; i < currentNode.documents.length; i++) {
-            stack.push(currentNode.documents[i]);
-          }
-        }
-
-        const nextState = await request.getDocumentList();
-        const stack2 = [...nextState];
-
-        while (stack2.length) {
-          const currentNode = stack2.pop();
-          if (results.includes(currentNode.id)) {
-            currentNode.isSpread = true;
-          }
-
-          const documents = [...currentNode.documents];
-          for (let i = 0; i < documents.length; i++) {
-            stack2.push(documents[i]);
-          }
-        }
+        const nextState = getRightNextState(
+          this.documentList.state,
+          await request.getDocumentList()
+        );
 
         this.documentList.setState(nextState);
       },
@@ -72,76 +46,21 @@ export default class App {
         history.pushState(null, null, `/${id}`);
         this.documentContent.setState(await request.getDocumentItem(id));
 
-        // 현재 데이터에서 isSpread가 true인 id를 results 배열에 넣는다.
-        const prevState = structuredClone(this.documentList.state);
-
-        const stack = prevState;
-        const results = [];
-        // 트리 순회
-        while (stack.length) {
-          const currentNode = stack.pop();
-
-          if (currentNode.isSpread) {
-            results.push(currentNode.id);
-          }
-
-          for (let i = 0; i < currentNode.documents.length; i++) {
-            stack.push(currentNode.documents[i]);
-          }
-        }
-
-        const nextState = await request.getDocumentList();
-        const stack2 = [...nextState];
-
-        while (stack2.length) {
-          const currentNode = stack2.pop();
-          if (results.includes(currentNode.id)) {
-            currentNode.isSpread = true;
-          }
-
-          const documents = [...currentNode.documents];
-          for (let i = 0; i < documents.length; i++) {
-            stack2.push(documents[i]);
-          }
-        }
+        const nextState = getRightNextState(
+          this.documentList.state,
+          await request.getDocumentList()
+        );
 
         this.documentList.setState(nextState);
       },
       onRemoveSubPageButtonClick: async (e) => {
         const id = e.currentTarget.id;
         await request.deleteDocumentItem(id);
-        // 현재 데이터에서 isSpread가 true인 id를 results 배열에 넣는다.
-        const prevState = structuredClone(this.documentList.state);
 
-        const stack = prevState;
-        const results = [];
-        // 트리 순회
-        while (stack.length) {
-          const currentNode = stack.pop();
-
-          if (currentNode.isSpread) {
-            results.push(currentNode.id);
-          }
-
-          for (let i = 0; i < currentNode.documents.length; i++) {
-            stack.push(currentNode.documents[i]);
-          }
-        }
-
-        const nextState = await request.getDocumentList();
-        const stack2 = [...nextState];
-
-        while (stack2.length) {
-          const currentNode = stack2.pop();
-          if (results.includes(currentNode.id)) {
-            currentNode.isSpread = true;
-          }
-
-          const documents = [...currentNode.documents];
-          for (let i = 0; i < documents.length; i++) {
-            stack2.push(documents[i]);
-          }
-        }
+        const nextState = getRightNextState(
+          this.documentList.state,
+          await request.getDocumentList()
+        );
 
         this.documentList.setState(nextState);
       },
@@ -180,40 +99,10 @@ export default class App {
         this.documentContent.timer = setTimeout(async () => {
           await request.updateDocumentItem(id, this.documentContent.state);
 
-          // 현재 데이터에서 isSpread가 true인 id를 results 배열에 넣는다.
-          const prevState = structuredClone(this.documentList.state);
-
-          const stack = prevState;
-          const results = [];
-          // 트리 순회
-          while (stack.length) {
-            const currentNode = stack.pop();
-
-            if (currentNode.isSpread) {
-              results.push(currentNode.id);
-            }
-
-            for (let i = 0; i < currentNode.documents.length; i++) {
-              stack.push(currentNode.documents[i]);
-            }
-          }
-
-          const nextState = await request.getDocumentList();
-          const stack2 = [...nextState];
-
-          while (stack2.length) {
-            const currentNode = stack2.pop();
-            if (results.includes(currentNode.id)) {
-              currentNode.isSpread = true;
-            }
-
-            const documents = [...currentNode.documents];
-            for (let i = 0; i < documents.length; i++) {
-              stack2.push(documents[i]);
-            }
-          }
-
-          this.documentList.setState(nextState);
+          const nextState = getRightNextState(
+            this.documentList.state,
+            await request.getDocumentList()
+          );
 
           this.documentList.setState(nextState);
         }, 1500);
